Consolidated multiple fields into single record
Changed the way the extension inserts data into Influx so that multiple measurements in a check all insert into the same table. For example, in a CPU check the guest, idle, system, user, etc fields all insert into a single record in accordance with Influx’s line protocol with multiple fields and associated measurments. Also changed one logger entry to debug so that every time a call is made to Influx it doesn’t log the response code.